Rename `matchcard` to `match`.
